level 5 added + mage/gunslinger finished

+++ Adventurer - Weapon Master Season 1/Assets/Script/Gameplay/Enemy/Enemy.cs
@@ -32,7 +32,7 @@ public class Enemy : MonoBehaviour
         currentHP -= damage; //health is decreased by damage dealt.
         EnemyHealthBar enemyHealthBar = healthBar.GetComponent<EnemyHealthBar>();
         enemyHealthBar.SetSize(PercentHP());
-        ShowFloatingText(damage);
+        ShowFloatingText((int)damage);
     }
     public void TakeDamageOverTime(float damagePerTick, float frequency, int numberOfTicks)
     {
